Make input file detection more robust / Now binary PitchTier file is supported
SLAM collects first all textgrid files by thier extension name (case is ignored) to proceed them one by one. For each textgrid file, it collects all source files which shares the same basename with it (case is ignored). Among these source files, it tries to find the best one on which it will
conduct the stylization task. SLAM will try to detect if there is a Praat PitchTier file it supports, if yes, it conducts the task on it. If there is no any PitchTier file it can recongine, it will tries to detect if there is an audio source files it can read, if yes, it estimates pitch on the
audio file by Swipe then applies stylisation on the filtered curve. if all tentative fails, SLAM switch to proceed another textgrid file until all textgrid files are analyzed in the way described above. About source files, we support for Praat PitchTier file, short text and binary (newly added) format; and for audio file, only wave format is possible for the moment.